Reimplement rotation and position listener components so that they track the absolute value:
normally, items we track are things such as the camera or hands, that by aframe documentation (e.g. https://aframe.io/docs/1.0.0/primitives/a-camera.html#manually-positioning-the-camera) are suggested to be enclosed in a camera rig. When one moves the rig, rather than the camera itself, the plain position attribute would not change.