First steps implementing a concept of ownership:
networked entities are now assigned to their creator. If multiple
actors try to create the same networked entity, the first actor will
be the owner, while the others will replace the local networked entity
with its remote representation.

An entity can be defined as persistent. In this case, whenever this
would become orphan (e.g. its owner leaves), this will be assigned to
a different online owner. The new owner will again replace the remote
object representation with the original networked entity.

The "ada-venue" example has been updated to show a situation where a
ball is controlled by the physics engine. Only the first peer
accessing the experience will apply physics to the ball. Subsequent
peers will receive object updates from the network. The state of the
ball will remain consistent as long as there is at least one peer in
the room.