Completely new approach for shared networked entities:
instead of replacing a local entity with a remote one on owneship
switch, simply interrupt network syncronization and notify the scene
of the event. Let the user decide what to do. In the ada-venue
example, we switch the physics on the shared ball from dynamic (local)
to kinematic (remote).

This approach works reliably at least on Oculus Quest and dramatically
reduces the number of objects we throw away every time.