WIP spawning of objects in the VR scene
when a filestorage is installed underneath an instance of this package, it is possible to spawn glb models as networked entities in the scene.
These entities can be set as volative (deleted when you leave), or permanent.
Currently, every model will be fitted to a sphere, but this should change. A lot of UI is also missing, for instance to list and manage existing entities.