New component "bound-to-entity", that will constrain an entity in the bounding box of another one
When the entity leaves the designated area, it will be forcefully repositioned to a specified spot.
In order for this to work, a change was needed in the physics engine so that the physics can be synced properly also on dynamic objects.

The change was also the chance to reduce object instantiation in other components.