Let go the awkward mediastream-listener component and use a different approach for ReadyPlayer me avatars to look at entities in the room:
one can specify a querySelector indicating the things the avatar should look at. When any of these entities is in the line of sight, the avatar will look at it.

The readyplayerme-avatar component was also made independent of any other component. The mediastream-listener usages have been cleaned up from the code.