Avatar spawning reform:
make it possible to configure explicitly if a networked entity should
be spawned immediately or not. For avatars and hands, delay the
spawning to the moment when we enter the immersive mode (this was
already the behavior for VR devices).

People accessing the "enter-vr" page will be in a sort of "spectator
mode" until they enter the immersive mode.

This makes the behavior consistent between browser and headsets.