Implement chat package integration:
when the chat package is available and a chat room id configured for the VR experience, the room will be rendered as part of the menu.

It is currently not supported to send messages while in immersive mode. This is due to the limited input support provided by the translation from HTML to mesh. Typing in VR is cumbersome anyway...