Init the data structures relevant for collision calculations when the body is added, rather than on the fly
This prevents cornercases where, if a body is deleted before it can
cause a collision (e.g. be the "body0" of a collision), there won't be
any collisionKey associated with it, which in turn will prevent
collisionend from firing and the collision map to be cleared.

In such case, no collision with the object will be fired anymore...