Networked scene: improve websocket behavior in case of reconnection
Instead of queuing messages until websocket is open, we just do not
send the message.

Whenever the websocket is open, either at first connection or at
reconnection, we send all entities that we have attached to the scene,
plus a full update of their tracked properties.

The idea is that, in case of loss of connectivity, as long as the
clients do not leave the scene, the experience can be resumed as soon
as the connection is restored.