Rework again the autoplay logics:
the environment is now responsible to handle autoplay according to expectation. We do not postpone putting the scene on the page anymore as this would force us to include potentially unnecessary javascript in common code.
A new convenience component has been provided to start autoplay sounds as soon as we click on the page.