Reform again how we extend the assets on the environment scene with those needed for our features:
using a Mutation Observer, we detect the exact moment when the scene is added to the document, and at that time exactly, we append the extra assets according to the feature flags.
The list of models was also reformed in order to be rendered also in assets format, as using ajax to load the model could have resulted in tricky race conditions.