Include a custom build of https://github.com/c-frame/aframe-gltf-model-plus
This is a port of the gltf model loader from Mozilla, which supports
Hubs models features such as lightmaps, animations and so on embedded
by the Spoke blender plugin in form of gltf extensions.

This enables to use Mozilla Hubs models unmodified in our experience
(with a couple caveats).

Fork of the component used here is located at
https://github.com/Elettrotecnica/aframe-gltf-model-plus/tree/elettrotecnica

Notable differences from the c-frame version:
- dependencies from Networked Aframe have been removed from the build,
  so components depending on this won't be available
- "LUTTonemapping", requiring a custom aframe build, is also disabled
  and falls back to "NoToneMapping"